Comments added to ObjectMapper class Abstract function are move to a separate class from ObjectMapperTest(JDBC related function)
