+++ .gitignore
@@ -4,3 +4,4 @@
 /dist
 .classpath
 .project
+.settings
