+++ src/com/cinergix/mapper/ObjectMapper.java
@@ -4,9 +4,36 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * This class provides main functionalities of “ResultObjectMapper”.
+ * <code>ObjectMapper</code> converts <code>ResultSet</code> data into java objects. To do this 
+ * conversion Java classes should be annotated with <code>ResultMapped</code> in class level. To 
+ * assign the column data to a property of a mapped object that property should be annotated with 
+ * <code>ResultField</code> and label of the column should be assigned to <code>fieldName</code> 
+ * property of <code>ResultField</code>.
+ * 
+ * Column value from  <code>ResultSet</code> can be manipulated before assign to property. To do 
+ * that, class should be annotated with<code>ResultTransformerClass</code> and the type of 
+ * transformer (Which should implement <code>IResultTransformer</code>) should be assigned to 
+ * <code>treasformerClass</code> property of <code>ResultTransformerClass</code>.
+ * 
+ * If a property is Custom type and that class also annotated as mentioned above then 
+ * <code>ResultSet</code> will be assigned recursively.
+ * @author rasekaran
+ *
+ * @param <T>
+ */
+
 public class ObjectMapper<T> {
 
-	public List<T> mapResultSetToObject( ResultSet result ) {
+	/**
+	 * Converts and returns a list of objects of type <code>T</code>. Those objects will have 
+	 * column values of given result of type <type>ResultSet</code>. 
+	 * @param result SQL result of type <code>ResultSet</code>.
+	 * @param dataClass Type of object in list which is returned by this method.
+	 * @return This method returns a list of object f type <code>T</code>
+	 */
+	public List<T> mapResultSetToObject( ResultSet result, Class<T> dataClass ) {
 		
 		if( result == null ){
